Exception handling - delete

+++ src/main/java/com/projetospbt/course/services/UserService.java
@@ -4,35 +4,44 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.projetospbt.course.entities.User;
 import com.projetospbt.course.repositories.UserRepository;
+import com.projetospbt.course.services.exceptions.DatabaseException;
 import com.projetospbt.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll(){
+
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
-	
+		 try {
+		        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+		        repository.deleteById(id);
+		    }catch (ResourceNotFoundException e){
+		        throw new ResourceNotFoundException(id);
+		    }catch (DataIntegrityViolationException e) {
+				throw new DatabaseException(e.getMessage());
+			}
+		}
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
@@ -43,7 +52,7 @@ public class UserService {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
-		
+
 	}
 
 }
